Only attempt to remove previously checkout out target if target folder exists

+++ checkout.ps1
@@ -6,7 +6,9 @@ param (
 
 [uri] $RepositorBaseUrl = "https://github.com/$RepositoryOwner/$RepositoryName.git"
 
-Remove-Item -Recurse -Verbose -Path $TargetFolder
+if (Test-Path $TargetFolder) {
+    Remove-Item -Recurse -Verbose -Path $TargetFolder
+}
 
 $git = Get-Command -CommandType Application "git" | Select-Object -First 1
 & "$git" clone "$RepositorBaseUrl" -- "$TargetFolder"
